+++ R/Likelihoods.R
@@ -26,17 +26,20 @@ logLikelihoodNull <- function(theta, logRr, seLogRr) {
   result <- 0
   sd <- 1/sqrt(theta[2])
   if (sd < 1e-6) {
+    # Note: not faster when vectorized
     for (i in 1:length(logRr)) {
       result <- result - dnorm(theta[1], logRr[i], seLogRr[i], log = TRUE)
     }
     
   } else {
+    # Note: not faster when vectorized
     for (i in 1:length(logRr)) {
       result <- result - log(gaussianProduct(logRr[i], theta[1], seLogRr[i], sd))
     }
   }
   if (length(result) == 0 || is.infinite(result))
     result <- 99999
+  # print(sprintf("theta = c(%s, %s), ll = %s", theta[1], theta[2], result))
   result
 }
 
@@ -56,7 +59,7 @@ minLogLikelihoodErrorModel <- function(theta, logRr, seLogRr, trueLogRr) {
       return(Inf)
     } else if (sd < 1e-6) {
       return(-dnorm(logRr[i], mean, seLogRr[i], log = TRUE))
-      } else {
+    } else {
       return(-log(gaussianProduct(logRr[i], mean, seLogRr[i], sd)))
     }
   }
@@ -76,4 +79,53 @@ minLogLikelihoodErrorModelLegacy <- function(theta, logRr, seLogRr, trueLogRr) {
   if (is.infinite(result) || is.na(result))
     result <- 99999
   result
+}
+
+profileProduct <- function(x, mu, sigma, llApproximationFunction, likelihoodApproximation) {
+  return(exp(dnorm(x, mu, sigma, log = TRUE) + 
+               llApproximationFunction(x = x, parameters = likelihoodApproximation)))
+}
+
+logLikelihoodNullNonNormalLl <- function(theta, llApproximationFunction, likelihoodApproximations) {
+  if (theta[2] <= 0)
+    return(99999)
+  result <- 0
+  sd <- 1/sqrt(theta[2])
+  if (sd < 1e-6) {
+    for (i in 1:length(likelihoodApproximations)) {
+      result <- result - llApproximationFunction(x = theta[1], parameters = likelihoodApproximations[[i]])
+    }
+  } else {
+    for (i in 1:length(likelihoodApproximations)) {
+      result <- result - log(integrate(f = profileProduct,
+                                       lower = theta[1] - 10 * sd,
+                                       upper = theta[1] + 10 * sd, 
+                                       mu = theta[1],
+                                       sigma = sd,
+                                       llApproximationFunction = llApproximationFunction,
+                                       likelihoodApproximation = likelihoodApproximations[[i]])$value)
+    }
+  }
+  if (length(result) == 0 || is.infinite(result))
+    result <- 99999
+  # print(sprintf("theta = c(%s, %s), ll = %s", theta[1], theta[2], result))
+  result
+}
+
+normalLlApproximaton <- function(x, parameters) {
+  dnorm(x, mean = parameters$logRr, sd = parameters$seLogRr, log = TRUE)
+}
+
+customLlApproximation <- function(x, parameters) {
+  return(((exp(parameters$gamma * (x - parameters$mu)))) * 
+           ((-(x - parameters$mu)^2)/(2 * parameters$sigma^2)))
+}
+
+skewNormalLlApproximation <- function(x, parameters) {
+  if (is.infinite(parameters$sigma)) {
+    return(rep(0, length(x)))
+  }
+  return(log(2) + 
+           dnorm(x, parameters$mu, parameters$sigma, log = TRUE) + 
+           pnorm(parameters$alpha * (x - parameters$mu), 0, parameters$sigma, log.p = TRUE))
 }